Move the extension methods to a new file.
